removed the visualization from the dataprocessor for single responsibility and better maintainability
